refactor: rename file and use constants
